Move pc_carry_weight and item_weight into common library
