+++ include/boe/item.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "boe/game_file_structures.hpp"
+
+short item_weight(const item_record_type& item);
